Refactored some other packages
