+++ src/main/kotlin/com/estapar/infrastructure/api/GarageWebhookController.kt
@@ -3,10 +3,10 @@ package com.estapar.infrastructure.api
 import com.estapar.domain.car.logging.entry.CarEntryService
 import com.estapar.domain.car.logging.exit.CarExitService
 import com.estapar.domain.car.park.attempt.ParkAttemptService
-import com.estapar.infrastructure.api.dto.GarageEntryEvent
-import com.estapar.infrastructure.api.dto.GarageExitEvent
-import com.estapar.infrastructure.api.dto.GarageParkedEvent
-import com.estapar.infrastructure.api.dto.GarageWebhookEvent
+import com.estapar.infrastructure.api.dto.request.GarageEntryEvent
+import com.estapar.infrastructure.api.dto.request.GarageExitEvent
+import com.estapar.infrastructure.api.dto.request.GarageParkedEvent
+import com.estapar.infrastructure.api.dto.request.GarageWebhookEvent
 import org.slf4j.LoggerFactory
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
