+++ src/main/kotlin/com/estapar/infrastructure/api/ParkingReportController.kt
@@ -1,7 +1,7 @@
 package com.estapar.infrastructure.api
 
 import com.estapar.domain.car.park.report.ParkingReportService
-import com.estapar.infrastructure.api.dto.ParkingReportResponse
+import com.estapar.infrastructure.api.dto.response.ParkingReportResponse
 import org.slf4j.LoggerFactory
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
