+++ src/main/kotlin/com/estapar/infrastructure/api/dto/request/GarageEventType.kt
@@ -1,4 +1,4 @@
-package com.estapar.infrastructure.api.dto
+package com.estapar.infrastructure.api.dto.request
 
 enum class GarageEventType {
     ENTRY,
